chore(console): remove unused v1 guides and mdx-components
