+++ packages/console/src/assets/docs/guides/m2m-general/README.mdx
@@ -1,7 +1,7 @@
 import InlineNotification from '@/ds-components/InlineNotification';
-import Steps from '@/mdx-components-v2/Steps';
-import Step from '@/mdx-components-v2/Step';
-import ApplicationCredentials from '@/mdx-components-v2/ApplicationCredentials';
+import Steps from '@/mdx-components/Steps';
+import Step from '@/mdx-components/Step';
+import ApplicationCredentials from '@/mdx-components/ApplicationCredentials';
 import EnableAdminAccess from './components/EnableAdminAccess';
 import EnableAdminAccessSrc from './assets/enable-admin-access.png';
 import AppIdentifierSrc from './assets/api-identifier.png';
