+++ packages/console/src/assets/docs/guides/native-android-java/README.mdx
@@ -1,7 +1,7 @@
-import UriInputField from '@/mdx-components-v2/UriInputField';
+import UriInputField from '@/mdx-components/UriInputField';
 import InlineNotification from '@/ds-components/InlineNotification';
-import Steps from '@/mdx-components-v2/Steps';
-import Step from '@/mdx-components-v2/Step';
+import Steps from '@/mdx-components/Steps';
+import Step from '@/mdx-components/Step';
 
 <Steps>
