+++ packages/console/src/assets/docs/guides/native-capacitor/README.mdx
@@ -1,6 +1,6 @@
-import UriInputField from '@/mdx-components-v2/UriInputField';
-import Steps from '@/mdx-components-v2/Steps';
-import Step from '@/mdx-components-v2/Step';
+import UriInputField from '@/mdx-components/UriInputField';
+import Steps from '@/mdx-components/Steps';
+import Step from '@/mdx-components/Step';
 
 import capaticorIos from './assets/capacitor-ios.webp';
 import logtoSignInPage from './assets/logto-sign-in-page.webp';
