+++ packages/console/src/assets/docs/guides/web-gpt-plugin/README.mdx
@@ -1,6 +1,6 @@
-import Steps from '@/mdx-components-v2/Steps';
-import Step from '@/mdx-components-v2/Step';
-import UriInputField from '@/mdx-components-v2/UriInputField';
+import Steps from '@/mdx-components/Steps';
+import Step from '@/mdx-components/Step';
+import UriInputField from '@/mdx-components/UriInputField';
 
 import AlwaysIssueRefreshToken from './components/AlwaysIssueRefreshToken';
 import AiPluginJson from './components/AiPluginJson';
