+++ packages/console/src/assets/docs/guides/web-outline/README.mdx
@@ -1,6 +1,6 @@
-import UriInputField from '@/mdx-components-v2/UriInputField';
-import Steps from '@/mdx-components-v2/Steps';
-import Step from '@/mdx-components-v2/Step';
+import UriInputField from '@/mdx-components/UriInputField';
+import Steps from '@/mdx-components/Steps';
+import Step from '@/mdx-components/Step';
 
 import logtoSignInExperience from './assets/logto-sign-in-experience.png';
 import outlineHome from './assets/outline-home.png';
