+++ packages/console/src/assets/docs/guides/web-remix/README.mdx
@@ -1,10 +1,10 @@
-import UriInputField from '@/mdx-components-v2/UriInputField';
+import UriInputField from '@/mdx-components/UriInputField';
 import Tabs from '@mdx/components/Tabs';
 import TabItem from '@mdx/components/TabItem';
 import InlineNotification from '@/ds-components/InlineNotification';
 import { buildIdGenerator } from '@logto/shared/universal';
-import Steps from '@/mdx-components-v2/Steps';
-import Step from '@/mdx-components-v2/Step';
+import Steps from '@/mdx-components/Steps';
+import Step from '@/mdx-components/Step';
 
 <Steps>
