--- packages/console/src/mdx-components-v2/UriInputField/index.module.scss
@@ -1,27 +1,0 @@
-@use '@/scss/underscore' as _;
-
-.wrapper {
-  display: flex;
-  align-items: flex-start;
-  position: relative;
-
-  .field {
-    flex: 1;
-
-    .multiTextInput {
-      flex: 1;
-    }
-  }
-
-  .saveButton {
-    flex-shrink: 0;
-    margin: _.unit(6) 0 0 _.unit(2);
-  }
-}
-
-/* Avoid using element selector directly in mdx components, as it will
- * affect all elements in the entire admin console app.
- */
-.form {
-  margin: _.unit(4) 0;
-}
