+++ packages/console/src/mdx-components/ApplicationCredentials/index.module.scss
