+++ packages/console/src/mdx-components/ApplicationCredentials/index.tsx
