+++ packages/console/src/mdx-components/Steps/index.module.scss
